refactor: put array in submodule of prelude
